// src/lib/api/generated/model/backupDashboardSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { BackupDashboardSummarySchema } from './backupDashboardSummarySchema';
import type { RestoreDashboardSummarySchema } from './restoreDashboardSummarySchema';

export interface BackupDashboardSchema {
  backups: BackupDashboardSummarySchema;
  restores?: RestoreDashboardSummarySchema | null;
}