// src/lib/api/generated/model/backupJobPageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { BackupJobProjectionSchema } from './backupJobProjectionSchema';

export interface BackupJobPageSchema {
  items: BackupJobProjectionSchema[];
  page: number;
  page_size: number;
  total: number;
}